Define user CRUD permissions

+++ src/main/java/com/tiago/ecommerce/user/UserService.java
@@ -2,11 +2,16 @@ package com.tiago.ecommerce.user;
 
 import com.tiago.ecommerce.role.Role;
 import com.tiago.ecommerce.role.RoleRepository;
+import com.tiago.ecommerce.utils.EncodingUtils;
+import com.tiago.ecommerce.utils.PrincipalUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import javax.management.relation.RoleNotFoundException;
 import java.util.List;
 import java.util.UUID;
 
@@ -20,8 +25,25 @@ public class UserService {
 
     private final String USER_NOT_FOUND = "User not found";
 
-    public User save(User user) {
-        return userRepository.save(user);
+    public ResponseEntity<Object> save(User user) throws RoleNotFoundException {
+        if(PrincipalUtils.isUser()){
+            return ResponseEntity.status(403).body("You don't have permission to create users");
+        }
+
+        Role role_user = roleRepository.findByNameIgnoreCase("role_user")
+                                        .orElseThrow(RoleNotFoundException::new);
+
+        user.getRoles().add(role_user);
+        role_user.getUsers().add(user);
+
+        userRepository.save(user);
+        roleRepository.save(role_user);
+
+        user.setPassword(EncodingUtils.encode(user.getPassword()));
+
+        return PrincipalUtils.isAdmin()
+                ? ResponseEntity.ok(user)
+                : ResponseEntity.ok(userToDto(user));
     }
 
     public ResponseEntity<Object> delete(UUID id) {
@@ -29,6 +51,10 @@ public class UserService {
 
         if(user == null) {
             return ResponseEntity.status(404).body(USER_NOT_FOUND);
+        }
+
+        if(transactionalActionNotAllowed(id)) {
+            return ResponseEntity.status(403).body("You don't have permission to delete this user");
         }
 
         userRepository.delete(user);
@@ -42,11 +68,15 @@ public class UserService {
             return ResponseEntity.status(404).body(USER_NOT_FOUND);
         }
 
-        return ResponseEntity.ok(user);
+        return PrincipalUtils.isAdmin()
+                ? ResponseEntity.ok(user)
+                : ResponseEntity.ok(userToDto(user));
     }
 
-    public List<User> getAll() {
-        return userRepository.findAll();
+    public List<?> getAll() {
+        return PrincipalUtils.isAdmin()
+                ? userRepository.findAll()
+                : userRepository.findAll().stream().map(this::userToDto).toList();
     }
 
     public ResponseEntity<Object> update(UUID id, UserDto updatedUserDto) {
@@ -54,6 +84,10 @@ public class UserService {
 
         if(user == null) {
             return ResponseEntity.status(404).body(USER_NOT_FOUND);
+        }
+
+        if(transactionalActionNotAllowed(id)) {
+            return ResponseEntity.status(403).body("You don't have permission to update this user");
         }
 
         BeanUtils.copyProperties(updatedUserDto, user);
@@ -80,5 +114,22 @@ public class UserService {
         roleRepository.save(role);
 
         return ResponseEntity.ok("Role assigned to user");
+    }
+
+    private UserDto userToDto(User user) {
+        return new UserDto(user.getUsername(), null);
+    }
+
+    private boolean transactionalActionNotAllowed(UUID id) {
+        return id != getPrincipalId() && !PrincipalUtils.isAdmin();
+    }
+
+    private UUID getPrincipalId() {
+        String principalUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        User principal = userRepository.findByUsernameIgnoreCase(principalUsername).orElse(null);
+
+        return principal == null
+                ? null
+                : principal.getId();
     }
 }
